backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
